% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{exportBackbone}
\alias{exportBackbone}
\title{Exports the network backbone as tsv file}
\usage{
exportBackbone(net, output_path = file.path(getwd(), "backbone.tsv"))
}
\arguments{
\item{net}{A R6 NPAModel object}

\item{output_path}{A string correspond to the tsv output filepath.
If not provided, ``backbone.tsv`` will be generated in working directory.}
}
\value{
Nothing is returned. File is created.
}
\description{
Exports the network backbone as tsv file
}
