% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLQ.R
\name{getLQ}
\alias{getLQ}
\title{Get precomputed matrix for fast NPA computations}
\usage{
getLQ(Model, verbose = TRUE, pattern.gene = "^EXP\\\\(")
}
\arguments{
\item{Model}{A R list object with slot model containing nodes and edges
data.frame}

\item{verbose}{A logical. Default is \code{TRUE}, messages will be displayed
during the function execution}

\item{pattern.gene}{A character vector. Regular expression for filtering
downstrean genes}
}
\value{
A R list object with computed metric like L2, L3 matrices,
Qbackbone, ...
}
\description{
Get precomputed matrix for fast NPA computations
}
\examples{
data(Hs__CFA__Apoptosis__0__0__1)
# Hs__CFA__Apoptosis__0__0__1 <- getLQ(
#    Hs__CFA__Apoptosis__0__0__1, verbose = TRUE)
}
