% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hs__CFA__Apoptosis__0__0__1.R
\docType{data}
\name{Hs__CFA__Apoptosis__0__0__1.rda}
\alias{Hs__CFA__Apoptosis__0__0__1.rda}
\title{Apoptosis Network for Homo sapiens (version 0.0.1)}
\description{
Apoptosis Network for Homo sapiens (version 0.0.1)
}
\keyword{data}
