% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hs__CPR__Cell_Cycle__0__0__1.R
\docType{data}
\name{Hs__CPR__Cell_Cycle__0__0__1.rda}
\alias{Hs__CPR__Cell_Cycle__0__0__1.rda}
\title{Cell Cycle network for Homo sapiens (version 0.0.1)}
\description{
Cell Cycle network for Homo sapiens (version 0.0.1)
}
\keyword{data}
