% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Hs__IPN__Neutrophil_Signaling__0__0__1.R
\docType{data}
\name{Hs__IPN__Neutrophil_Signaling__0__0__1.rda}
\alias{Hs__IPN__Neutrophil_Signaling__0__0__1.rda}
\title{Neutrophil Signaling network for Homo sapiens (version 0.0.1)}
\description{
Neutrophil Signaling network for Homo sapiens (version 0.0.1)
}
\keyword{data}
