% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_data}
\alias{list_data}
\title{Retrieve the list of rda files}
\usage{
list_data(pattern = "*.rda")
}
\arguments{
\item{pattern}{A regex value for listing the rda files}
}
\value{
Vector of strings, list of data files without extension
}
\description{
Retrieve the list of rda files
}
\examples{
l <- list_data()

}
