% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_families}
\alias{list_families}
\title{Retrieves the list of available families}
\usage{
list_families(species = "Hs")
}
\arguments{
\item{species}{A string, species name. Default is `Hs` for Homo sapiens}
}
\value{
Vector of strings, the list of families
}
\description{
Retrieves the list of available families
}
\examples{
f <- list_families(species = "Mm")
}
