% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_models}
\alias{list_models}
\title{Retrieves the list of models}
\usage{
list_models(species = "Hs", family = NULL)
}
\arguments{
\item{species}{A string, species name}

\item{family}{A string, the familiy name, if `NULL`, all available models
are listed.}
}
\value{
Vector of strings, the list of model names
}
\description{
Retrieves the list of models
}
\examples{
f <- list_models(species = "Mm", family = "CPR")
}
