% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_model.R
\name{load_models}
\alias{load_models}
\title{Retrieves the list of NPAModels R6 class objects for a
given species}
\usage{
load_models(species = "Hs", verbose = FALSE)
}
\arguments{
\item{species}{A string, species name.}

\item{verbose}{A logical, if TRUE, messages are shonw in the
console.}
}
\value{
A list object with NPAModel R6 object per slots.
}
\description{
Retrieves the list of NPAModels R6 class objects for a
given species
}
\examples{
list_models(species = "Mm")
# ms <- load_models(species = "Mm")
}
