% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minimizeNetwork.R
\name{minimizeNetwork}
\alias{minimizeNetwork}
\title{Reduces the slots of a network object to the minimum
required slots for getLQ function to be applied}
\usage{
minimizeNetwork(model)
}
\arguments{
\item{model}{List with few slots to be cleanded}
}
\value{
Same object with some slots removed
}
\description{
Reduces the slots of a network object to the minimum
required slots for getLQ function to be applied
}
