% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mm__CPR__Cell_Cycle__0__0__1.R
\docType{data}
\name{Mm__CPR__Cell_Cycle__0__0__1.rda}
\alias{Mm__CPR__Cell_Cycle__0__0__1.rda}
\title{Cell Cycle network for Mus musculus (version 0.0.1)}
\description{
Cell Cycle network for Mus musculus (version 0.0.1)
}
\keyword{data}
