% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mm__CPR__Jak_Stat__0__0__1.R
\docType{data}
\name{Mm__CPR__Jak_Stat__0__0__1.rda}
\alias{Mm__CPR__Jak_Stat__0__0__1.rda}
\title{Jak Stat network for Homo sapiens (version 0.0.1)}
\description{
Jak Stat network for Homo sapiens (version 0.0.1)
}
\keyword{data}
