% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/Mm__IPN__Epithelial_Innate_Immune_Activation__0__0__1.R
\docType{data}
\name{Mm__IPN__Epithelial_Innate_Immune_Activation__0__0__1.rda}
\alias{Mm__IPN__Epithelial_Innate_Immune_Activation__0__0__1.rda}
\title{Epithelial Innate Immune Activation network for Mus musculus
(version 0.0.1)}
\description{
Epithelial Innate Immune Activation network for Mus musculus
(version 0.0.1)
}
\keyword{data}
