% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mm__TRA__ECM_Degradation__0__0__1.R
\docType{data}
\name{Mm__TRA__ECM_Degradation__0__0__1.rda}
\alias{Mm__TRA__ECM_Degradation__0__0__1.rda}
\title{ECM Degradation network for Mus musculus (version 0.0.1)}
\description{
ECM Degradation network for Mus musculus (version 0.0.1)
}
\keyword{data}
