% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NPAModel.R
\docType{class}
\name{NPAModel}
\alias{NPAModel}
\title{Class providing object with methods for handling a network model}
\format{\code{\link{R6Class}} object.}
\usage{
NPAModel
}
\value{
Object of \code{\link{R6Class}} named NPAModel containing
network model
}
\description{
Class providing object with methods for handling a network model
}
\section{Fields}{

\describe{
\item{\code{data}}{A R list object containing model data}

\item{\code{species}}{A string, species name}

\item{\code{family}}{A string, family name}

\item{\code{version}}{A string, version of the model}
}}

\keyword{datasets}
