% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessNetwork.R
\name{preprocessNetworks}
\alias{preprocessNetworks}
\title{Preprocess network data.}
\usage{
preprocessNetworks(minimize = FALSE, lq = TRUE)
}
\arguments{
\item{minimize}{Logical, all unecessary slots are removed}

\item{lq}{Logical, prepare the network for NPA computation}
}
\value{
A NULL value.
}
\description{
Preprocess network data.
}
\examples{
## This example processes the package models to make them minimal
preprocessNetworks(minimize = TRUE, lq = FALSE)
}
